Transforming Video stream

+++ src/video-stream/video-stream.service.ts
@@ -1,16 +1,36 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { spawn, ChildProcess } from 'child_process';
-import { Observable, Subject } from 'rxjs';
+import { Observable, Subject, ReplaySubject } from 'rxjs';
 import { VideoStreamConfig } from './video-stream-config.model';
+import * as stream from 'stream';
+import * as StreamSplit from 'stream-split';
+import { of, concat } from 'rxjs';
+
+// TODO: separate file
+// Packet from data ( default binary = true )
+interface Packet {
+  data: Buffer | string;
+  binary: boolean;
+}
+
+interface capHeader {
+  lastIdrFrame: Buffer;
+  sps: Buffer;
+  pps: Buffer;
+}
 
 @Injectable()
 export class VideoStreamService {
   
+  // general
   private command = 'raspivid';
   private logger = new Logger(this.constructor.name);
   
+  // capture
   private capProcess: ChildProcess;
-  private capSubject: Subject<Buffer>;
+  private capSubject: Subject<Packet>;
+  private capHeader: capHeader = { lastIdrFrame: undefined, sps: undefined, pps: undefined};
+  private capStartingHeaderSubject: ReplaySubject<Packet>;
   
   // TODO: support live reload, read from file etc
   private config: VideoStreamConfig = {
@@ -20,18 +40,21 @@ export class VideoStreamService {
     framerate: 20,
   };
   
-  getCapture(): Observable<Buffer> {
-    if (this.isCapturing()) {
-      return this.capSubject;
+  getCapture(): Observable<Packet> {
+    if (!this.isCapturing()) {
+      this._startCapture();
     }
     
-    this._startCapture();
-    return this.capSubject;
+    return concat(this._getStreamHeader(), this.capSubject);
   }
   
   killCapture() {
     this.logger.log('Killing capture process');
-    this.capProcess.kill();
+    if (this.capProcess) {
+      this.capProcess.kill();
+    } else {
+      this.logger.log('Capture process is already stopped');
+    }
   }
   
   isCapturing(): boolean {
@@ -42,10 +65,11 @@ export class VideoStreamService {
     this.logger.log('Starting capture process');
     this._initState();
     this._setupEvents();
+    this._setupStream();
   }
   
   private _initState() {
-    this.capSubject = new Subject<Buffer>();
+    this.capSubject = new Subject<Packet>();
     this.capProcess = spawn(this.command, [
       '--width', `${this.config.width}`,
       '--height', `${this.config.height}`,
@@ -54,14 +78,65 @@ export class VideoStreamService {
       '--timeout', '0',
       '-o', '-',
     ]);
+    this.capStartingHeaderSubject = new ReplaySubject<Packet>();
+  }
+  
+  private _setupStream() {
+    const NalSeparator = new Buffer([0,0,0,1]);
+    const self = this;
+  
+    this.capProcess.stdout
+      .pipe(new StreamSplit(NalSeparator))
+      .pipe(new stream.Transform({ transform: function (chunk, encoding, callback) {
+          const completeChunk = Buffer.concat([NalSeparator, chunk]);
+          const chunkType = chunk[0] & 0b11111;
+
+          switch(chunkType) {
+            case 7: // SPS
+              if (!self.capHeader.sps) {
+                self._updateInitialState(completeChunk);
+              }
+              self.capHeader.sps = completeChunk;
+              break;
+            case 8: // PPS
+              if (!self.capHeader.pps) {
+                self._updateInitialState(completeChunk);
+              }
+              self.capHeader.pps = completeChunk;
+              break;
+            case 5: // IDR
+              if (!self.capHeader.lastIdrFrame) {
+                self._updateInitialState(completeChunk);
+              }
+              self.capHeader.lastIdrFrame = completeChunk;
+              // @Fallthrough
+            default:
+              this.push(completeChunk);
+          }
+          
+          self._tryCompleteInitialState();
+
+          callback();
+      }})).on('data', data => {
+        this.capSubject.next({ data, binary: true });
+      });
+  }
+  
+  private _updateInitialState(chunk) {
+    if (this.capStartingHeaderSubject !== undefined) {
+      this.capStartingHeaderSubject.next({ data: chunk, binary: true });
+    }
+  }
+  
+  private _tryCompleteInitialState() {
+    if (this.capStartingHeaderSubject !== undefined &&
+        Object.values(this.capHeader).every(val => val != undefined)) {
+      this.capStartingHeaderSubject.complete();
+      this.capStartingHeaderSubject = undefined;
+    }
   }
   
   private _setupEvents() {
-    // receive and forward video stream
-    this.capProcess.stdout.on('data', data => {
-      this.capSubject.next(data);
-    });
-    
     // closing & exiting
     this.capProcess.on('close', code => {
       const message = `${this.command} closed all stdio with code ${code}`
@@ -97,5 +172,32 @@ export class VideoStreamService {
   private _cleanup() {
     this.capProcess = undefined;
     this.capSubject = undefined;
+    this.capHeader = { lastIdrFrame: undefined, sps: undefined, pps: undefined};
+    this.capStartingHeaderSubject = undefined;
+  }
+  
+  private _getStreamHeader(): Observable<Packet> {
+    const staticHeader: Packet = { 
+          data: JSON.stringify({
+            action: 'init',
+            width: this.config.width,
+            height: this.config.height
+          }),
+          binary: false
+        };
+        
+    return concat(of(staticHeader), 
+                  this.capStartingHeaderSubject ?
+                    this.capStartingHeaderSubject :
+                    of({
+                        data: this.capHeader.sps, 
+                        binary: true
+                      }, {
+                        data: this.capHeader.pps, 
+                        binary: true
+                      }, {
+                        data: this.capHeader.lastIdrFrame, 
+                        binary: true
+                      }));
   }
 }
